@@ -1,8 +1,8 @@
-import type { DocsPage } from "$lib/types.js";
+import type { DocsPage, DocsTopic } from "$lib/types.js";
 
 /** @type {import('./$types').LayoutServerLoad} */
 export async function load({ fetch }) {
   const response = await fetch("/api/docs");
-  const pages: DocsPage[] = await response.json();
-  return { pages };
+  const menuItems: (DocsPage | DocsTopic)[] = await response.json();
+  return { menuItems };
 }
